QSDK: DP:wifi: ath12k: Add reinject handler for ppe-ds flow
Add support for reinjecting frames in the direct switch
transmit completion path.

FW exception always occurs for the initial few sdwf packets.
In the case of DS, since the FW doesn't get the svc id,
it reinjects the packet back to the host, which then enqueues
it to TCL with all necessary sdwf fields populated.

Change-Id: I9ce84d6faf342aaab77933f8665ecb5e82484fd6
Signed-off-by: Prateek Sharma <prsha@qti.qualcomm.com>