UPSTREAM:PROTOCOL:cfg80211:mac80211:wifi: Support MLO with AFC
cfg80211 and mac80211:
Currently, the 6 GHz power mode is stored at an interface level within
the wdev structure. This does not work for multi-link devices where each
6 GHz link can have a different power mode.

To fix this, store the 6 GHz power mode in the link structure.
Add code changes across mac80211 and cfg80211 to support this.

ath11k:
Currently, for filling the set TPC WMI command, ath11k uses the
power mode from the wdev. But, this field is removed and
moved to a links field of the wdev in the latest cfg80211.

To fix this, use the power mode from the bss configuration instead
of directly from the wdev.

ath12k:
Currently, the driver does not have the support for handling
power mode change commands for an MLD with two 6 GHz links,
where each link is in different 6 GHz power modes.

Also, for filling the set TPC WMI command, ath12k uses the
power_type field from the 'ieee80211_bss_conf' structure for
STA interface and the 'reg_6g_power_mode' from the wdev for
AP type intarface. But, this field in the wdev is removed and
moved to a links field of the wdev in the latest cfg80211.
To fix this,
   (1) Add new attribute link id to the power mode change command.
   (2) Use the link id sent by the application as part of the
       power mode change command to determine the link / radio
       for which the power mode change command is sent.
   (3) Also, pass the link id as part of the power mode changed
       event to the userspace to indicate which link has
       changed its power mode.
   (4) Use the power mode from the 'ieee80211_bss_conf' structure
       for both STA and AP interfaces. This works because, in
       the latest cfg80211, the power mode is updated in the
       'ieee80211_bss_conf' structure for both STA and AP
       interfaces.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch

Change-Id: I477e93afceb662a4b32b730adc616ace1ae4c177
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>