Upstream:DP:wifi: ath12k: Configure header length while  enabling monitor length
Currently rx_hdr_len is not configured and as a default option 256bytes
of hardware header is getting passed from monitor destination ring for
each MPDUs. But host needs only 64 bytes to construct the packet header.
So reducing the rx header size from 256 bytes to 64 bytes to avoid the
extra header subscription which helps to reduce the CPU cycles while
processing this TLV.

Patch-dependency : None

Change-Id: I1e8ecf58f13cde3d2dbdcb8c886e0a88b4eacae3
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>