QSDK:CP: wifi: mac80211: set 5GHz bw_limit to 320MHz for EHT to enable 240MHz support
EHT connections on the 5GHz band currently do not set a bandwidth limit,
which may prevent the use of 240MHz operation when puncturing is enabled.
The default behavior does not explicitly allow 240MHz, as the bw_limit
remains unset for 5GHz EHT.

Also, the parsing logic for QCN vendor-specific elements incorrectly
casts the pointer when assigning the 240MHz capability structure, and the
conditional block for parsing QCN elements is misplaced, preventing it from
being executed in some cases when EHT 240MHz is enabled.

Set the bw_limit to IEEE80211_CONN_BW_LIMIT_320 for 5GHz EHT connections
to enable 240MHz operation. This ensures that the connection logic
correctly reflects the capabilities of EHT in the 5GHz band. Fix the
pointer cast in the QCN vendor IE parser and move the parsing block outside
the WMM condition to ensure it is always evaluated when applicable.

Change-Id: I6cae65fb4cb37f93982612a123fcc958f52d5ab1
Signed-off-by: Balamurugan Ramar <bramar@qti.qualcomm.com>
Signed-off-by: Ramasamy Kaliappan <rkaliapp@qti.qualcomm.com>
Signed-off-by: Maharaja Kennadyrajan <mkenna@qti.qualcomm.com>