QSDK:PLATFORM:wifi:ath12k: Fix qdss memory allocation failure
Currently for WKK qdss_mem allocation is getting failed as
Host doesn't have the necessary memeory. This is because
for PCI and AHB chipsets DMA memory will be used so the
AFC memory is not allocated from DDR, but thesize of AFC
region was accounted in DDR because of which There's no
sufficient memory left for QDSS.

Fix this by not accounting AFC memory region type as part
of HOST DDR. Also remove the sanity checks and paddr
initialization as it is not necessary.

Change-Id: Ic0eea7b7df143c4d90e2f53b09901fce7e11179b
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>