WAR:CP:wifi:ath12k Remove WARN_ON() in dec_num_stations
In 13.0 mac_dec_num_stations never had WARN_ON to align with
that remove the WARN_ON in 13.1 as well.

As part of ath12k_mac_peer_cleanup_all() we make ar->num_stations
as 0 and later in disassociation flow when mac_dec_num_stations() is called
this warning will pop up.

TODO: Check if this initialization to 0 can be removed in ath12k_mac_peer_cleanup_all()

Change-Id: Ib350de62fc73edd599e8ed002e0eb02f5f207a5f
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>