UPSTREAM: PROTOCOL: wifi: ath12k: Fix regd update for weather radar
Currently, when regd is updated for weather radar band, the
index for tracking the filled reg rules is not incremented.
However, the successive rules are filled in the successive
array indices. But, the number of reg rules in the regd is
set to the number of rules filled in the regd.

Since this is not incremented, the variable indicating the
number of reg rules in the regd and the number of reg rules
actually filled in the regd will not match.

This causes cfg80211 to reject the regd update.

Fix this by incrementing the index for tracking the filled
reg rules when the weather radar band is updated.
Also, ensure that the number of reg rules in the regd
is set correctly based on the actual number of rules filled.

Fixes: 7a510b184414 ("UPSTREAM:PROTOCOL:wifi:ath12k: AFC patch rebase")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch

Change-Id: Iaaedc6a2ee6755987a034ad8a06f02a0e7413f46
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>