UPSTREAM:PROTOCOL:cfg80211:mac80211: Fix STA disconnection with AFC
Currently, For 6 GHz, when a STA processes the channel switch announcement
(CSA) frame, the target channel is determined from the CSA IE
and the 6 GHz power mode is read from the beacon's 6 GHz operation IE's
control field. The STA validates if the target channel is valid for the
power mode from the beacon. If the target channel is not valid for the
power mode, the STA disconnects from the AP.

When the AP switches between channels of different power
mode, the STA side validation fails as the target channel may not be
valid for the power mode in the beacon. This results in the STA
disconnection from the AP.

Since , we do not know the power mode of the target channel from the CSA
IEs, to fix this, verify whether the target channel is valid in at least
one of the power modes, then allow the STA to switch to the target
channel in the calculated power mode.

After switching to the new channel with the calculated (or temporary)
power mode, the STA receives the Root AP's power mode from the first
beacon. If the power mode in the beacon differs from the calculated
(or temporary) power mode, the STA switches to the new mode.
However, if the beacon's power mode is not supported by the STA, the
STA is brought down.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
subsys/0340-UPSTREAM-PROTOCOL-cfg80211-mac80211-Store-6-GHz-powe.patch

Change-Id: I1e51c61a70c11cd5205b45d3d9848972eb6f76b8
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>