mac80211: Apply radio mask changes while the  network interface is running
In case of 2 GHz + 5 GHz MLO, the 5G 40 MHz scan failed because the
radio mask was not set due to netif running check in mac80211,
which led to the VAP failing to come up.

Add changes to apply radio mask even if netif is running as dynamic link
add/remove is supported.

Add changes to set and get radio mask in ucode. Previously, ucode sets
the radio mask for the first interface created in ucode and not set for
all the radios. Due to this, trigger scan fails with radio mask check in
mac80211 and it cause MLO vap failure. Add changes to set radio mask
for all other radios from the ucode.

Change-Id: Ib6edbd7995e94fdb6c7450fcb77dca122235e520
Signed-off-by: Balamurugan Ramar <bramar@qti.qualcomm.com>