QSDK:CP:wifi:ath12k: fix NULL ar access in ath12k_mac_op_change_vif_links()
In ath12k_mac_op_change_vif_links(), when we are trying to access
arvif->ar during non-recovery cases, it may be NULL if VDEV is not
created already and arvif->ar is not assigned in
ath12k_mac_vdev_create() yet.

Fix this with arvif->ar check to prevent NULL access exception.

Fixes: Ia3217d85c04 ("QSDK:CP:wifi:ath12k: Fix double free of skb")

Change-Id: I5616bd21082f7c3c2b04b7f3d9460fe68e13c325
Signed-off-by: Sathishkumar Muruganandam <murugana@qti.qualcomm.com>