UPSTREAM:PROTOCOL:wifi:ath12: Rectify 5.9 GHz AP bring up issue
An AP that does not have the 5.9 GHz BDF service flag enabled is able
to operate on the 5.9 GHz channels.

Some changes present in the 3.1 component's
625-ath12k-supporting-5.9G-band-with-160Mhz.patch
were missed in the 4.0 component's
0317-ath12k-supporting-5.9G-band-with-160Mhz.patch during the package
upgrade. These changes involve disabling the 5.9 GHz channels in the
sband, depending on whether the 5.9 GHz BDF service flag is set.

To fix this issue, incorporate the missing changes into the 4.0 component.

Tag: REGULATORY
Fixes: I026a851ab13('ath12k: add support to select all 6 GHZ regulatory rules')

Change-Id: I71a28847fd70d4ade07d502898ffdef24c3f6a22
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>