QSDK:PROTCOL: wifi: cfg80211/mac80211: fix UBPR template update
Currently whenever beacon is updated, UBPR changed BSS flag is set always
irrespective of whether interval is getting changed or not. This leads to
failure in CSA when count is 0.

Fix this issue by comparing the interval value first and proceeding with
changed flag setting only when interval changes.

While at it, remove setting the flag during CSA and CCA handler explicitly.
The helper API already decided and sets the flag hence there is no need to
set it again at the caller.

Fixes: f652dc967619 ("package upgrade: mac80211 patch set 11")
Fixes: 6e20df2d6aef ("wifi: mac80211: pkg upgrade changes set 17")
Signed-off-by: Aditya Kumar Singh <aditya.kumar.singh@oss.qualcomm.com>
Change-Id: I2d88801159bd99c7acad3a044cf5b66c92292c5c