QSDK:DP:wifi:ath:ath12k: Add support to program tid for active MSCS session
Add support to get mscs priority for peer having an MSCS session
based on MSCS context established during handshake.
Using the wifi plugins, NSS modules check with the WLAN driver
on what priority to use for the skb before accelerating the
UL and DL flows.
Provide the correct priority to be set in the skb based on MSCS
context.

In case of DL, check if skb->priority is having non-zero tid
and then decide if HLOS tid override has to be done

Tag: MSCS

Change-Id: I114d6cbf9b66149c29823e4371761d28c232783f
Signed-off-by: Ramanathan Choodamani <rchoodam@qti.qualcomm.com>
Signed-off-by: Vivek Chettri <vchettri@qti.qualcomm.com>