UPSTREAM:CP:wifi:ath12k:Fix vdev id extraction during CU update
CU update is taken care of all the radios and all the VAPs configured in
the radio. To extract the vdev id, allocated vdev map is OR'ed with bit
for the range of 0 - 16. But in case of split phy, the vdev is span from
0-16*number of radios. This causes loop to exit without checking vdev id
17 - 31 for another radio. When highest index 16 is returned, iterator
might not find the arvif which causes below warning to get flooded:

[123862.309682] ath12k_wifi7_pci 0001:01:00.0: No VIF found for vdev 17
[123862.426378] ath12k_wifi7_pci 0001:01:00.0: No VIF found for vdev 16

Add a new variable to keep track of maximum vdev supported in SoC which
will help to loop through all the vdev created.

Patch-work:None
Patch-dependency:
0292-0003-ath12k-Handle-CU-and-BPCC-for-probe-and-assoc-respon.patch

Change-Id: I839319c7ce9c115b0775a6ab851c4cc86eed0f70
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>