UPSTREAM:CP: wifi:ath11k: Avoid ath11k radio access during regulatory channel update
During the regulatory channel update scenario, radio does not have
proper pdevs allocated during that instance. So, during this channel
update when we try to take the ath11k radio, it goes for crash as
there is no NULL check by returning error value.

This ath11k radio is not required for channel update, this is taken
for checking the vdev type. Even in the working case everytime ath11k
radio returns invalid vdev type which is not required for channel
update.

Channel update is happening in the default regulatory of ath11k base
everytime and this vdev type is not required for regulatory channel
update.

Change-Id: I9d16ec0cae1d0f2ad0713810d8ed08a090877618
Signed-off-by: Gautham Kumar Senthilkumaran <gauthamk@qti.qualcomm.com>