UPSTREAM:PROTOCOL:wifi-scripts:Enable Mirrored SCS protocol by default
Add support to set enable_mscs to true by default during AP bring up.
With this, the AP will advertise MSCS capabilities in it's beacon,
assoc and probe response frames only.
The actual session is established only when a STA sends an MSCS
request and the AP accepts it.

Patch-dependency: None
Tag: MSCS

Change-Id: If048e1a923725a81d7110a930f2113c4ca864740
Signed-off-by: Ramanathan Choodamani <rchoodam@qti.qualcomm.com>