UPSTREAM:DP:wifi:ath12k: Add API for ath12k_dp_peer search using address and ieee80211_sta
In the following scenarios, hostapd might not send NL command for delete
station to driver:

1. In congested environment with multiple clients, de-authentication
   frames from client might be missed in AP Rx

2. In roaming scenario, client can roam across the link and it is
   not mandatory for client to send the deauthentication frame to AP

3. Sometimes client may go to out of coverage and try to connect back
   with another radio with same mac address

In such cases, when NL command for add station is received in ath12k
driver without delete station command being received, ath12k_dp_peer
creation fails since there is an existing entry in driver with the
same mac address and this results in station connection failure.

This issue is introduced due to introduction of ath12k_dp_peer as part
of Unified Driver framework.

In order to address this, search for ath12k_dp_peer using mac address
and ieee80211_sta in creation and deletion paths which makes it unique.

Fixes: 70b5a1a98 (wifi: ath12k: Modularization of peer object Part-2)

Patch Dependency:
0116-0001-wifi-ath12k-Define-ath12k_dp_peer-structure-APIs-for.patch
0116-0002-wifi-ath12k-Attach-and-detach-ath12k_dp_link_peer-to.patch

Change-Id: Ie7dbc657542dfebae9f244e298629ce6f8ddee21
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>