QSDK:wifi-scripts:WAR: Bring Monitor interfaces down first during wifi down.
Currently in case of Monitor + AP when CAC is in progress and wifi down
is exercised in remove_link flow the corresponding link_conf will be
assigned to NULL and later a WARN_ON will be seen on the same link_conf
being NULL in ath12k_mac_change_chanctx_cnt_iter(),
ath12k_mac_change_chanctx_fill_iter(),

In the exisiting flow this was not the case as the Monitor iface has
Exclusive context and the ieee80211_recalc_radar_chanctx() is avoided
based on the number of ifaces are assigned the same channel context.

With latest upstream design monitor will be sharing the same context and
it will be assigned even before CAC is completed. Because of which the
above flow is exercised and Warnings are seen.

Change the sequence of wifi down to so that Monitor will be brought down
first to avoid this Warnings.

Change-Id: I454d1e4ce04d0c168605cb23dba4c2c067421522
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>