QSDK:DP:wifi:ath12k: Remove unused rx tlv words
removed the unused rx tlv fields to reduce
the cache misses which will improve the KPI performance.

in support of fse meta data use in rx path there
some additional words added in the rx_tlv which is
causing the KPI decrement.
As part fixing this we have identify there
is two Dwords which can remove from regular
tlv subscription.
with this change we got ~800Mbps improvement in TCP UL.

Change-Id: I3b9bfa761419e485ec41cc24855e0162d88035e3
Signed-off-by: KARTHIK KUMAR T <kartt@qti.qualcomm.com>