UPSTREAM:PROTOCOL:wifi:ath12k: Do not add link id for non-MLD interfaces
Currently, when power mode is changed in the ath12k driver, the driver
sends the event 'QCA_NL80211_VENDOR_SUBCMD_POWER_MODE_CHANGE_COMPLETED'
to the userspace along with the new power mode and the link id of the
interface. This is done for both MLO and non-MLO interfaces. However,
sending the link id for non-MLO interfaces is not necessary and can
cause issues in userspace handling.

To fix this, modify the ath12k driver to include the link id only for
MLD interfaces when sending the power mode change event.

Fixes: I477e93afceb ("UPSTREAM:PROTOCOL:cfg80211:mac80211:wifi: Support MLO with AFC")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch

Change-Id: Ia0b968aa71239161c9cd9f492015807b58d97cce
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>