UPSTREAM:CP: wifi:ath12k: Use correct lock for ahvif list access
Fix the lock used when accessing the RCU-protected ahvif link list.
Instead of relying on the ath12k hardware mutex, which is not suitable
for this purpose, use the wiphy mutex from mac80211, which provides the
appropriate protection for safely checking the ahvif list.

Patch-dependency: 0339-0001-wifi-ath12k-refactor-primary-link-selection-logic.patch
Patch-work: None

Change-Id: I056c10bde6ec1acb404cbf9fec01c0d90f1a02fe
Signed-off-by: Gautham Kumar Senthilkumaran <gauthamk@qti.qualcomm.com>