UPSTREAM:PROTOCOL: wifi: cfg80211: mac80211: ath12k: Add 6 GHz client type support
Currently, the default client type for 6 GHz is set to regular client.
Using regular client may be suboptimal for few regulatory domains
where the Tx powers of regular client is lower than that of a
subordinate client.

Add a new field to the regulatory domain structure to indicate
the client types supported for the regulatory domain. This is
a bitmask field that can indicate the various supported client
types.

The underlying hardware may indicate the supported client types
for the regulatory domain as part of the new regd field.

Add support in cfg80211 for determining the client type for 6 GHz
based on the regulatory domain and the 6 GHz AP power mode.
Note that the subordinate client is valid only for LPI power mode.
Use regular client type for 6 GHz APs if the regulatory domain
does not indicate the supported client types.

Add support in mac80211 to calculate the client power mode based on
the client type instead of using the default regular client type.

Add the 6 GHz client type received from the Firmware to the wiphy
regulatory domain. Also, advertise the support
for regular client in the wiphy regd irrespective of the client type
received from the firmware. This is to ensure that the regular client,
which is the default client type, is always available for use.

Also, parse the TPE IE only for the corresponding client type.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
subsys/0340-UPSTREAM-PROTOCOL-cfg80211-mac80211-Store-6-GHz-powe.patch

Change-Id: I05ac8bf2a91073ffe831b9aea81facc519c4d0af
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>