UPSTREAM:PROTOCOL: wifi: mac80211: Continue CAC due to punctured channel radar
Currenty, when a radar is detected, mac80211 cancels the CAC timer
(if running) and marks the radar affected sub-channel(s) as unavailable.

Consider a case where we have started radar detection on a DFS channel
with a user configured puncture pattern. If we detect a radar on
the punctured subchannel, we cancel the running CAC timer and mark
the subchannel as unavailable.

We may, however, still continue the radar detection on the DFS
channel. This is valid because the radar is detected on a punctured
subchannel and the punctured channel will not be used for transmission.

Fix this by allowing the CAC timer to continue when a radar is detected
on a punctured subchannel. Make sure to mark / update the state of the
subchannel as unavailable.

Fixes: Ib960576bdd7 ("DFS interference bitmap support during DFS interference detection.")
Patch-work: None
Tag: DFS
Patch-dependency:
subsys/0084-mac80211-radar-bitmap-during-DFS-radar-detection.patch

Change-Id: Id57d810d0c301cb6f228a804bddffc36be289dd6
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>