QSDK:PLATFORM:wifi:ath12k: Advertise recovery Mode  after Q6 recovery.
In case of multipd devices when host is subscribed to userpd crash
notification and after successful Mode-0 recovery the userpd will not
able to retain the previous recovery Mode because of which on the
next FW assert the crash notification will not be receieved for all
the userpd's.

Fix this issue by advertising the recovery Mode every time after
recovery.

Change-Id: I5c93add8799ee90fa67ff933d6a212acbd325b4c
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>