QSDK:CP: mac80211: Fix Quilt directory does not exist error for extns
This is the fix for wifi7 directory doesnot exist error
for quilt prepare command when QCN_EXTN is enabled.

Quilt copies all the patches to temp patches directory in the backports
directory. These pacthes are not yet applied. Then copies the
extension source files into respective qcn_extns folders in the
backports.

During this step, quilt tries to copy the files from ath/wifi7
directory into the backport's ath12k/wifi7/qcn_extns directory. And
it encounters no directory exist error for wifi7. This happens because
the wifi7 direcory is created during applying patches stage. But in
quilt prepare, the patches are applied by the user in the later
stages.

Hence, Add fix to create wifi7 directory to avoid this error
for quilt prepare command.

This change should not be there when the frameworks changes adding wifi7 is
downstreamed through package upgrade.

Fixes: I5008dd3fda7e29f43a1784fcb9b7e5293098b289 ("QSDK:CP: mac80211: makefile: Add support to build minimal patches extension files")

Change-Id: Ia1bff366e89344c8baa5b7ed3478607a920d9732
Signed-off-by: Sri Vidya Gunturi <sgunturi@qti.qualcomm.com>