QSDK:PROTOCOL: mac80211.sh: Enable SCS by default without UCI file update
SCS is supported and enabled by default. For enabling SCS by default we
update the SCS enable config in hostapd config file and also update the
UCI config file with scs_enable '1'.

Since SCS is expected to be enabled OOB, we just enable the config in
the hostapd config file and we dont update the UCI every time for
default
configuration.

If the SCS config is overridden in the UCI file, the same would also be
reflected in the hostapd config file.

Change-Id: I5a5fde5341c374078a19b9f7fcc126cc8ce96bcb
Signed-off-by: Vivek Chettri <vchettri@qti.qualcomm.com>