QSDK:DP:wifi:ath12k:Remove stats_disable check from TLV parsing function
Always parse HTT_PPDU_STATS_TAG_USR_COMMON in TLV parsing function by
removing stats_disable check as it does not allow to perform stats
processing. This function is responsible for parsing
important TLV fields and assigning the struct for post-processing.
The check was incorrectly preventing proper parsing of critical TLVs

This is an existing change in 13.0 (added by
982-02-UPSTREAM-CP-DP-wifi-ath12k-Add-telemetry-stats-for-QCA-.patch). Porting the same

Fixes: 98811828c60d ("add telemetry and vendor support for ErP and energy services")

Patch-dependency:
0130-15-UPSTREAM-CP-wifi-ath12k-Use-HTT_PPDU_STATS-for-tx-completion-sta.patch

Change-Id: I04530ccdf513d165232e7928dc2d0451e7ff349b
Signed-off-by: Arpitha Ramesha <arpirame@qti.qualcomm.com>